possibly fixed float rounding

+++ dev_ws/src/rdr/src/motor_driver.cpp
@@ -25,11 +25,11 @@ class MotorDriver : public rclcpp::Node
   private:
   int servoMin = 220;
   int servoMax = 520;
-    static int map ( int x, int in_min, int in_max, int out_min, int out_max) {
-        int toReturn =  (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min ;
+    static int map ( float x, int in_min, int in_max, int out_min, int out_max) {
+        float toReturn =  (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min ;
         // For debugging:
         // printf("MAPPED %d to: %d\n", x, toReturn);
-        return toReturn ;
+        return (int)toReturn;
     }
 
     void topic_callback(const geometry_msgs::msg::Twist::SharedPtr msg) const
